% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitdc-package.R
\docType{package}
\name{fitdc}
\alias{fitdc}
\alias{fitdc-package}
\title{fitdc: R package for decoding FIT files.}
\description{
A pure R FIT file decoder. After having a package pulled from CRAN due to
licensing issues with the FIT SDK, the only way to bring that package back
from the dead was to put this together. As far as I'm concerned, it is fit
for purpose, but I am aware it is not a \emph{complete} implementation. If
this does not meet your needs in its current state, feel free to submit a
patch.
}
\details{
Anyhow, performance is suprisingly good, and the package is written to have
no external dependencies. I hope you find it useful!
}

